[Fix] fix error of implicit conversion
- This commit fixes the error of implicit conversion.
- This commit aims to fix the following error:
```
../test/unittest/unittest_nntrainer_save_with_dtype.cpp:682:61: error: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648 [-Werror,-Wimplicit-const-int-float-conversion]
    input.setValue(0, 0, 0, w, static_cast<float>(rand()) / RAND_MAX - 0.5f);
                                                          ~ ^~~~~~~~
/usr/include/stdlib.h:87:18: note: expanded from macro 'RAND_MAX'
```

+++ test/unittest/unittest_nntrainer_save_with_dtype.cpp
@@ -679,7 +679,9 @@ static nntrainer::Tensor buildInput(unsigned int width,
   nntrainer::Tensor input(dim);
   srand(seed);
   for (unsigned int w = 0; w < width; ++w)
-    input.setValue(0, 0, 0, w, static_cast<float>(rand()) / RAND_MAX - 0.5f);
+    input.setValue(0, 0, 0, w,
+                   static_cast<float>(rand()) / static_cast<float>(RAND_MAX) -
+                     0.5f);
   return input;
 }
